General refactor all controllers, form errors (java side)

+++ src/main/java/it/univaq/disim/mwt/j2etpapp/presentation/rest/NotificationRestController.java
@@ -4,13 +4,12 @@ import it.univaq.disim.mwt.j2etpapp.business.BusinessException;
 import it.univaq.disim.mwt.j2etpapp.business.NotificationBO;
 import it.univaq.disim.mwt.j2etpapp.domain.NotificationClass;
 import it.univaq.disim.mwt.j2etpapp.domain.UserClass;
-import it.univaq.disim.mwt.j2etpapp.security.UserDetailsImpl;
 import it.univaq.disim.mwt.j2etpapp.utils.JSONDealer;
+import it.univaq.disim.mwt.j2etpapp.utils.UtilsClass;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,20 +23,26 @@ public class NotificationRestController {
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity getAllNotifications() throws BusinessException {
-        UserClass principal = (SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof UserDetailsImpl) ? ((UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser() : null;
+        UserClass principal = UtilsClass.getPrincipal();
 
-        List<NotificationClass> notifications = notificationBO.findByUserTargetId(principal.getId());
+        if(principal != null) {
+            List<NotificationClass> notifications = notificationBO.findByUserTargetId(principal.getId());
+            return new ResponseEntity(JSONDealer.ObjectToJSON(notifications), HttpStatus.OK);
+        }
 
-        return new ResponseEntity(JSONDealer.ObjectToJSON(notifications), HttpStatus.OK);
+        return new ResponseEntity(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping("delete_all")
     public ResponseEntity deleteAllNotifications() {
-        UserClass principal = (SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof UserDetailsImpl) ? ((UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser() : null;
+        UserClass principal = UtilsClass.getPrincipal();
 
-        notificationBO.deleteAllByUserTargetId(principal.getId());
+        if(principal != null) {
+            notificationBO.deleteAllByUserTargetId(principal.getId());
+            return new ResponseEntity(HttpStatus.OK);
+        }
 
-        return new ResponseEntity(HttpStatus.OK);
+        return new ResponseEntity(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping("{notificationId}/delete")
@@ -49,11 +54,14 @@ public class NotificationRestController {
 
     @PostMapping("read_all")
     public ResponseEntity readAllNotifications() {
-        UserClass principal = (SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof UserDetailsImpl) ? ((UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser() : null;
+        UserClass principal = UtilsClass.getPrincipal();
 
-        notificationBO.readAllByUserTargetId(principal.getId());
+        if(principal != null) {
+            notificationBO.readAllByUserTargetId(principal.getId());
+            return new ResponseEntity(HttpStatus.OK);
+        }
 
-        return new ResponseEntity(HttpStatus.OK);
+        return new ResponseEntity(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping("{notificationId}/read")
